Litt opprydding - fjern spor av tps
