+++ src/main/kotlin/no/nav/familie/ef/søknad/api/dto/Søkerinfo.kt
@@ -1,7 +1,7 @@
 package no.nav.familie.ef.søknad.api.dto
 
-import no.nav.familie.ef.søknad.api.dto.tps.Barn
-import no.nav.familie.ef.søknad.api.dto.tps.Person
+import no.nav.familie.ef.søknad.api.dto.pdl.Barn
+import no.nav.familie.ef.søknad.api.dto.pdl.Person
 
 data class Søkerinfo(val søker: Person,
                      val barn: List<Barn>) {
