+++ src/main/kotlin/no/nav/familie/ef/søknad/api/dto/pdl/Barn.kt
@@ -1,0 +1,12 @@
+package no.nav.familie.ef.søknad.api.dto.pdl
+
+import java.time.LocalDate
+
+data class Barn(val fnr: String,
+                val navn: String,
+                val alder: Int,
+                val fødselsdato: LocalDate,
+                val harSammeAdresse: Boolean,
+                val annenForelder: AnnenForelder?)
+
+
