+++ src/main/kotlin/no/nav/familie/ef/søknad/api/dto/søknadsdialog/Adresse.kt
@@ -4,4 +4,3 @@ data class Adresse(val adresse: String,
                    val postnummer: String,
                    val poststed: String?)
 
-// TODO mangler mapping av data fra ui? : adressetillegg? Kommune?
